Move static home page components into static_components folder.
