+++ client/src/components/static_components/Loading.js
