+++ client/src/components/static_components/SuggestEdit.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./../css/About.css";
+import "./../../css/About.css";
 
 const SuggestEdit = () => (
   <main className="main-container">
